Move inputs component into shared folder
