+++ src/logged_in/components/GraphEditor/NodeModal.js
@@ -2,7 +2,7 @@
 import {FieldArray, Form, Formik} from "formik";
 import React from "react";
 import PropTypes from 'prop-types';
-import FormikTextField from '../inputs/FormikTextField'
+import FormikTextField from '../../../shared/inputs/FormikTextField'
 import { Button, Grid, Slide, DialogContent, DialogTitle, Dialog} from '@material-ui/core'
 
 import * as yup from 'yup'
