+++ src/logged_in/components/NewGrabitCard.js
@@ -7,7 +7,7 @@ import {
 import {Form, Formik} from "formik";
 import {useMutation} from "@apollo/react-hooks";
 import {CREATE_QUERY} from "../../shared/costants/queries";
-import FormikTextField from "./inputs/FormikTextField";
+import FormikTextField from "../../shared/inputs/FormikTextField";
 
 const useStyles = makeStyles({
         card: {
