+++ src/shared/inputs/FormikTextField.js
@@ -5,9 +5,7 @@ import {TextField} from "@material-ui/core";
 const FormikTextField = (props) => {
     const [field, meta] = useField(props);
     const errorText = meta.error && meta.touched ? meta.error : "";
-    return <TextField {...field} {...props}
-                      helperText={errorText}
-                      error={!!errorText}/>;
+    return <TextField {...field} {...props} helperText={errorText} error={!!errorText}/>;
 };
 
 export default FormikTextField;
